+++ apps/showcase/src/app/app.component.html
@@ -3,7 +3,7 @@
     <range-selector class="layout-main-range"></range-selector>
     <fhir-chart class="layout-main-chart" width="100%" height="100%" [floatingContent]="legend"> </fhir-chart>
     <ng-template #legend>
-      <dataset-tags-legend></dataset-tags-legend>
+      <fhir-chart-tags-legend></fhir-chart-tags-legend>
       <fhir-chart-legend></fhir-chart-legend>
     </ng-template>
     <fhir-chart-summary class="layout-main-summary"></fhir-chart-summary>
