+++ libs/ngx-charts-on-fhir/src/index.ts
@@ -16,8 +16,8 @@ export * from './lib/fhir-chart-summary/home-measurement-summary.service';
 export * from './lib/fhir-chart-layout/fhir-chart-layout.module';
 export * from './lib/fhir-chart-layout/fhir-chart-layout.component';
 export * from './lib/fhir-chart-legend/fhir-chart-legend.module';
-export * from './lib/fhir-chart-legend/dataset-tags-legend/dataset-tags.service';
-export * from './lib/fhir-chart-legend/dataset-tags-legend/dataset-tags-legend.component';
+export * from './lib/fhir-chart-legend/fhir-chart-tags-legend/fhir-chart-tags.service';
+export * from './lib/fhir-chart-legend/fhir-chart-tags-legend/fhir-chart-tags-legend.component';
 export * from './lib/dynamic-table/dynamic-table.component';
 export * from './lib/dynamic-table/dynamic-table.module';
 export * from './lib/fhir-data/fhir-data.service';
