--- libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/dataset-tags.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { DatasetTagsService } from './dataset-tags-legend/dataset-tags.service';
-
-describe('DatasetTagsService', () => {
-  let service: DatasetTagsService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(DatasetTagsService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
