+++ libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/fhir-chart-legend.module.ts
@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FhirChartLegendComponent } from './fhir-chart-legend.component';
 import { MatIconModule } from '@angular/material/icon';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { MatButtonModule } from '@angular/material/button';
-import { SortDatasets } from './sort-datasets.pipe';
+import { SortDatasets } from './fhir-chart-legend/sort-datasets.pipe';
+import { FhirChartLegendComponent } from './fhir-chart-legend/fhir-chart-legend.component';
 import { FhirChartLegendItemComponent } from './fhir-chart-legend-item/fhir-chart-legend-item.component';
-import { DatasetTagsLegendComponent } from './dataset-tags-legend/dataset-tags-legend.component';
+import { FhirChartTagsLegendComponent } from './fhir-chart-tags-legend/fhir-chart-tags-legend.component';
 
 @NgModule({
-  declarations: [FhirChartLegendComponent, SortDatasets, FhirChartLegendItemComponent, DatasetTagsLegendComponent],
+  declarations: [FhirChartLegendComponent, SortDatasets, FhirChartLegendItemComponent, FhirChartTagsLegendComponent],
   imports: [CommonModule, MatIconModule, MatButtonModule, OverlayModule],
-  exports: [FhirChartLegendComponent, DatasetTagsLegendComponent],
+  exports: [FhirChartLegendComponent, FhirChartTagsLegendComponent],
 })
 export class FhirChartLegendModule {}
