+++ libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/fhir-chart-legend/fhir-chart-legend.component.css
