+++ libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/fhir-chart-legend/sort-datasets.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { Dataset } from '../data-layer/data-layer';
-import { byMostRecentValue } from '../fhir-chart-summary/statistics.service';
+import { Dataset } from '../../data-layer/data-layer';
+import { byMostRecentValue } from '../../fhir-chart-summary/statistics.service';
 
 @Pipe({ name: 'sortDatasets' })
 export class SortDatasets implements PipeTransform {
