+++ libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/fhir-chart-tags-legend/fhir-chart-tags-legend.component.html
