+++ libs/ngx-charts-on-fhir/src/lib/fhir-chart-legend/fhir-chart-tags-legend/fhir-chart-tags.service.ts
@@ -15,7 +15,7 @@ type TagStyles = Record<string, DatasetOptions>;
 @Injectable({
   providedIn: 'root',
 })
-export class DatasetTagsService {
+export class FhirChartTagsService {
   private tagStyles = new BehaviorSubject<TagStyles>({
     Home: {
       pointStyle: 'triangle',
